requirements.txt
fastapi
uvicorn[standard]
cassandra-driver
